pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rtclient"
version = "0.5.3"
description = "A client for the RT API"
authors = ["Microsoft Corporation"]

[tool.poetry.dependencies]
python = ">=3.10"
aiohttp = "3.10.9"
azure-identity = "*"
pydantic = "2.9.2"
aiohappyeyeballs = "2.4.3"
aiohttp-retry = "2.8.3"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anyio = "4.6.0"
async-timeout = "4.0.3"
apscheduler = "3.11.0"
attrs = "24.2.0"
certifi = "2024.8.30"
charset-normalizer = "3.3.2"
click = "8.1.7"
exceptiongroup = "1.2.2"
fastapi = "0.115.0"
frozenlist = "1.4.1"
h11 = "0.14.0"
idna = "3.10"
multidict = "6.1.0"
pydantic-core = "2.23.4"
pyjwt = "2.10.1"
python-dotenv = "1.0.1"
requests = "2.32.3"
sniffio = "1.3.1"
starlette = "0.38.6"
twilio = "9.4.1"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
uvicorn = "0.31.1"
websockets = "13.1"
yarl = "1.14.0"
azure-cosmos = "4.9.0"
python_multipart="0.0.20"

[tool.poetry.dev-dependencies]
ruff = "*"
black = "*"
python-dotenv = "*"
soundfile = "*"
numpy = "*"
scipy = "*"
pytest = "*"
pytest-asyncio = "*"

[tool.poetry.scripts]


[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
extend-ignore = ["UP007"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
